pyproject.toml
[build-system]
requires = ['setuptools>=40.8.0', 'wheel']
build-backend = 'setuptools.build_meta:__legacy__'

[project]
name = "django-vnpay"
version = "0.0.6"
description = "A quick package for integrate vnpay payment"
readme = "README.md"
authors = [{ name = "Son Lam", email = "son.lg@eoh.io" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "django >= 4.0.0",
    "djangorestframework >= 3.12.4",
    "django-ipware >= 4.0.0",
]
keywords = ["django", "vnpay"]
requires-python = ">=3.10"

[project.urls]
Homepage = "https://github.com/laughlamz/django-vnpay"
